Many final changes, some tweaks to be made at home
Tweaks to be made at home, just styling stuff and whatnot. Otherwise, students, course, and custom order is done, faculty idk whats happening / is it worth going over? And mockup is just a couple of fixes, but good to show, and textbooks 101 is still a mess, lets see what we can get done.